requirements.txt
git+https://github.com/openai/whisper.git
python-dotenv
gradio -q 
openai
transformers
googletrans==3.1.0a0
